# backend/api-zod/src/generated/types/invoiceInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GST Billing & Inventory Management API
 * OpenAPI spec version: 0.1.0
 */
import type { InvoiceInputRoundingOption } from "./invoiceInputRoundingOption";
import type { InvoiceItem } from "./invoiceItem";

export interface InvoiceInput {
  invoiceDate: string;
  customerId?: number;
  customerName: string;
  customerAddress: string;
  customerGstin?: string;
  items: InvoiceItem[];
  totalQuantity: number;
  subTotal: number;
  totalCgst: number;
  totalSgst: number;
  grandTotal: number;
  roundingOption: InvoiceInputRoundingOption;
  customRounding?: number;
  finalAmount: number;
  amountInWords: string;
}